refactor(domain): pure Validate + drop FlowField.Unique
Validate now takes pre-resolved FlowResolvedFields instead of loading
the schema itself — no ctx, no QueryExecutor, no I/O. Single source of
truth: whatever rules Resolve surfaces on FlowField.Validation are
exactly what Validate enforces. Validation lives in its own file/test
pair.

FlowField.Unique and FlowFieldUniqueScope are dropped: nothing
consumed them, and uniqueness is a storage concern owned by
domain.AttributeUniqueness (per ADR 008), not a flow-field render
trait.
